Lift query fold processing to the shared query module and make implicit time ops optional
This is still pretty ugly, but is trying to preserve old behavior as much as
possible. We introduce two new parameters to `argvQuery`:

- `implicitTimeOps`: if true, enable all of our implicit time operations.
` -doFolds`: if true, do fold processing (which used to be in `coronerList`)
  uniquely).

Also add support for the count aggregate.
