+++ lib/cli/query.js
@@ -88,7 +88,7 @@ function parseFilterFlags(filter) {
   return flags;
 }
 
-function argvQuery(argv) {
+function argvQueryPrefold(argv, implicitTimestampOps) {
   var query = {};
   var d_age = null;
 
@@ -154,7 +154,7 @@ function argvQuery(argv) {
     query.order = argv.sort.map(parseSortTerm);
   }
 
-  if (!query.filter[0].timestamp)
+  if (!query.filter[0].timestamp && implicitTimestampOps)
     query.filter[0].timestamp = [];
 
   if (argv.time) {
@@ -228,7 +228,7 @@ function argvQuery(argv) {
     } else {
       query.select = [ argv.select ];
     }
-  } else if (argv.table === 'objects') {
+  } else if (argv.table === 'objects' && implicitTimestampOps) {
     query.fold = {
       'timestamp' : [['range'], ['bin']]
     };
@@ -268,7 +268,7 @@ function argvQuery(argv) {
     d_age = '1M';
   }
 
-  if (d_age) {
+  if (d_age && implicitTimestampOps) {
     var now = Date.now();
     var target = parseInt(now / 1000) - timeCli.timespecToSeconds(d_age);
     var oldest = Math.floor(target);
@@ -280,7 +280,7 @@ function argvQuery(argv) {
     range_start = oldest;
     range_stop = Math.floor(now / 1000);
 
-    if (query.fold && query.fold.timestamp) {
+    if (query.fold && query.fold.timestamp && implicitTimestampOps) {
       var ft = query.fold.timestamp;
       var i;
 
@@ -293,12 +293,91 @@ function argvQuery(argv) {
   }
 
   if (argv.table === 'objects') {
-    if (!query.filter[0].timestamp || query.filter[0].timestamp.length === 0)
+    if (!query.filter[0].timestamp || query.filter[0].timestamp.length === 0 &&
+      implicitTimestampOps)
       query.filter[0].timestamp.push([ 'greater-than', 0 ]);
   }
 
   return { query: query, age: d_age };
 }
+
+/*
+ * Generate a query from argv.
+ *
+ * if implicitTimestampOps is false, don't add anything to do with timestamp.
+ * this is used by alerts among other things to allow for the user to enter
+ * queries without having to know the JSON syntax.
+ */
+function argvQuery(argv, implicitTimeOps=false, doFolds=false) {
+  const { query, age } = argvQueryPrefold(argv, implicitTimeOps);
+
+  if (!doFolds) {
+    return { query, age };
+  }
+
+  /*
+   * This was originally lifted from morgue.js and not refactored because it's
+   * fragile.
+   */
+  function fold(query, attribute, label) {
+    var argv, i;
+
+    if (!query.fold)
+      query.fold = {};
+
+    if (Array.isArray(attribute) === false) {
+      attribute = [ attribute ];
+    }
+
+    for (i = 0; i < attribute.length; i++) {
+      var modifiers, j;
+
+      modifiers = attribute[i].split(',');
+      argv = modifiers[0];
+      modifiers.shift();
+
+      for (j = 0; j < modifiers.length; j++) {
+        modifiers[j] = parseInt(modifiers[j]);
+        if (isNaN(modifiers[j]) === true) {
+          errx('Modifiers must be integers.');
+        }
+      }
+
+      if (!query.fold[argv])
+        query.fold[argv] = [];
+
+      query.fold[argv].push([label].concat(modifiers));
+    }
+  }
+
+  const folds = [
+    [argv.last, 'last'],
+    [argv.first, 'first'],
+    [argv.tail, 'tail'],
+    [argv.head, 'head'],
+    [argv.object, 'object'],
+    [argv.histogram, 'histogram'],
+    [argv.distribution, 'distribution'],
+    [argv.unique, 'unique'],
+    [argv.mean, 'mean'],
+    [argv.sum, 'sum'],
+    [argv.quantize, 'bin'],
+    [argv.bin, 'bin'],
+    [argv.range, 'range'],
+    [argv.count, 'count'],
+  ];
+
+  /* Apply requested folds to query */
+  folds.forEach(function(attr_op) {
+    const [attr, op] = attr_op;
+    if (attr)
+      fold(query, attr, op);
+  });
+
+  return { query, age };
+}
+
+
 
 module.exports = {
   argvQuery,
